added flags and gc

+++ pysitemap/__init__.py
@@ -5,7 +5,7 @@ import signal
 from pysitemap.base_crawler import Crawler
 
 
-def crawler(root_url, out_file, out_format='xml', maxtasks=100, batch_size=10000, prefix=None):
+def crawler(root_url, out_file, out_format='xml', maxtasks=100, batch_size=10000, prefix=None, load=False):
     """
     run crowler
     :param root_url: Site root url
@@ -17,11 +17,13 @@ def crawler(root_url, out_file, out_format='xml', maxtasks=100, batch_size=10000
     loop = asyncio.get_event_loop()
 
     # TODO (noahfang) - put name in assets file and use here and in base_crawler.py
-    try:
-      os.remove('seen.sqlite')
-      os.remove('todo.sqlite')
-    except:
-      pass
+    # If we do not want to load existing data, delete the data
+    if not load:
+        try:
+            os.remove('seen.sqlite')
+            os.remove('todo.sqlite')
+        except:
+            pass
 
     c = Crawler(root_url, out_file=out_file, out_format=out_format, maxtasks=maxtasks, batch_size=batch_size, prefix=prefix)
     loop.run_until_complete(c.run())
